pyproject.toml
[tool.poetry]
name = "heisman"
version = "0.1.0"
description = ""
authors = ["Alexander Talbott <atalbott13@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
jupyter = "^1.0.0"
pandas = "^1.4.4"
sportsipy = "^0.6.0"
black = "^22.8.0"
mypy = "^0.971"
cfbd = "^4.4.4"
psycopg2 = "^2.9.3"
matplotlib = "^3.5.3"
scikit-learn = "^1.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
